refactor: move and rename calculateEdges
